Fix all golangci-lint issues and Docker build
- Update default image to aerospike:ce-8.1.1.1 (official Docker Hub)
- Fix 25 golangci-lint issues (modernize, unparam, unused, prealloc, etc.)
- Replace interface{} with any across codebase
- Add enterprise image detection for new :ee- tag format
- Fix .dockerignore to include go:embed resources (scripts/)
- Add --platform=$BUILDPLATFORM to Dockerfile for multi-arch builds
- Switch lint CI to make lint for custom plugin support
- Add pre-commit config with standard hooks and golangci-lint

+++ internal/storage/pvc.go
@@ -108,7 +108,7 @@ func extractOrdinal(pvcName, stsName string) (int32, bool) {
 	// Parse ordinal.
 	var ordinal int32
 	for _, c := range pvcName[idx+1:] {
-		ordinal = ordinal*10 + int32(c-'0')
+		ordinal = ordinal*10 + (c - '0')
 	}
 
 	return ordinal, true
